Add Bamboozled::API:Tabular
The BambooHR API allows you to add rows to the following tables: jobInfo, employmentStatus, compensation, dependentsemergencyContacts or (contacts), and custom tables (customTableName). To make this possible in Bamboozled we need a Tabular class that takes the table name, columns, and employee id. Note: the #generate_xml class is duplicating code from the employee class but I want to wait to extract this into a module or the Base class.